Moved misplaced container files into components/. Removed unused files
