--- app/components/Tab.css
@@ -1,23 +1,0 @@
-.tab {
-  top: 8%;
-  height: 3%;
-  width: 100%;
-  background-color: #f5f5f5;
-  border-color: #bfbfbf;
-  border-style: solid;
-  border-width: 1px 0px;
-  position: fixed !important;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.arrow {
-  opacity: 0.5;
-}
-
-.tabContent {
-  font-size: 13px;
-}
